Extract CDN to own file
